+++ src/main/java/com/epam/esm/model/entity/UserOrder.java
@@ -17,12 +17,12 @@ import java.time.Instant;
 public class UserOrder {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private Double cost;
     @Column(name="create_date")
     private Instant createDate;
     @Column(name="user_id")
-    private int userId;
+    private Long userId;
     @Column(name="certificate_with_tag_id")
-    private int CertificateWithTagId;
+    private Long CertificateWithTagId;
 }
