+++ src/main/java/com/epam/esm/repository/CertificateRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CertificateRepository extends JpaRepository<Certificate, Integer> {
+public interface CertificateRepository extends JpaRepository<Certificate, Long> {
     List<Certificate> findByNameContaining(String pattern);
 
     List<Certificate> findByDescriptionContaining(String pattern);
