+++ src/main/java/com/epam/esm/security/JwtIssuer.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class JwtIssuer {
 
     private final JwtProperties properties;
-    public String issue(int userId, String name, List<String> roles){
+    public String issue(Long userId, String name, List<String> roles){
         return JWT.create()
                 .withSubject(String.valueOf(userId))
                 .withExpiresAt(Instant.now().plus(Duration.of(1, ChronoUnit.DAYS)))
